remove plain http implementation
