+++ lib/blob-server/index.js
@@ -1,65 +1,90 @@
-// @ts-check
-import http from 'node:http'
 import assert from 'node:assert'
+import fp from 'fastify-plugin'
 import { filetypemime } from 'magic-bytes.js'
+
 import { SUPPORTED_BLOB_VARIANTS } from '../blob-store/index.js'
 
-/** @typedef {import('../blob-store').BlobId} BlobId */
-/** @typedef {Error & { statusCode: number }} HttpError */
+export default fp(blobServerPlugin, {
+  fastify: '4.x',
+  name: 'mapeo-blob-server',
+})
 
-export class BlobServer {
-  #server
-  #blobStore
+/**
+ * @typedef {Object} BlobsServerPluginOpts
+ *
+ * @property {import('../blob-store/index.js').BlobStore} blobStore
+ * @property {string} [prefix]
+ */
 
-  /**
-   * @param {object} opts
-   * @param {import('../blob-store').BlobStore} opts.blobStore
-   */
-  constructor({ blobStore }) {
-    this.#blobStore = blobStore
-    this.#server = http.createServer(this.requestHandler)
-  }
+/**
+ * @typedef {Object} ParsedParams
+ *
+ * @property {string} driveId
+ * @property {import('../types').BlobType} type
+ * @property {import('../types').BlobVariant<ParsedParams['type']>} variant
+ * @property {string} name
+ */
 
-  /**
-   * Start the blob server listening for requests
-   *
-   * @param {number} [port] Defaults to `0` e.g. randomly selected port
-   * @param {string} [hostname] Network interface to listen on, defaults to 127.0.0.1 (e.g. only local)
-   */
-  async listen(port = 0, hostname = '127.0.0.1') {
-    return listenPromise(this.#server, port, hostname)
-  }
+/** @type {import('fastify').FastifySchema['params']} */
+const PARAMS_JSON_SCHEMA = /** @type {const} */ {
+  type: 'object',
+  required: ['driveId', 'type', 'variant', 'name'],
+  properties: {
+    driveId: { type: 'string' },
+    type: { type: 'string', enum: Object.keys(SUPPORTED_BLOB_VARIANTS) },
+    variant: {
+      type: 'string',
+      enum: [...new Set(Object.values(SUPPORTED_BLOB_VARIANTS).flat())],
+    },
+    name: { type: 'string' },
+  },
+}
 
-  /**
-   * Stop the blob server from listening for new requests (existing requests will complete)
-   *
-   * @returns {Promise<void>}
-   */
-  async close() {
-    return new Promise((resolve) => {
-      this.#server.close(() => {
-        resolve()
-      })
-    })
-  }
+/** @type {import('fastify').FastifyPluginAsync<BlobsServerPluginOpts>} */
+async function blobServerPlugin(fastify, options) {
+  // We call register here so that the `prefix` option can work if desired
+  // https://fastify.dev/docs/latest/Reference/Routes#route-prefixing-and-fastify-plugin
+  fastify.register(routes, options)
+}
 
-  /** @type {import('node:http').RequestListener} */
-  async requestHandler(req, res) {
-    try {
-      assert(req.method === 'GET', 'Invalid Request')
-      const blobId = parseRequestParams(req)
+/** @type {import('fastify').FastifyPluginAsync<Omit<BlobsServerPluginOpts, 'prefix'>>} */
+async function routes(fastify, options) {
+  const { blobStore } = options
 
-      const entry = await this.#blobStore.entry(blobId)
+  fastify.get(
+    '/:driveId/:type/:variant/:name',
+    { schema: { params: PARAMS_JSON_SCHEMA } },
+    /**
+     * @param {import('fastify').FastifyRequest<{ Params: ParsedParams }>} request
+     */
+    async (request, reply) => {
+      const { driveId, type, variant, name } = request.params
+
+      const blobId = constructBlobId(driveId, type, variant, name)
+
+      const entry = await blobStore.entry(blobId)
+
+      if (!entry) {
+        reply.code(404)
+        throw new Error('Entry not found')
+      }
+
       const { blob, metadata } = entry.value
 
-      const drive = this.#blobStore.getDrive(blobId.driveId)
+      const drive = blobStore.getDrive(blobId.driveId)
       const blobs = await drive.getBlobs()
 
+      if (!blobs) {
+        reply.code(500)
+        throw new Error('Hyperdrive instance has no blobs instance')
+      }
+
+      // TODO: Need to update Hyperdrive to latest and then pass `wait: false` here
       const blobsStream = blobs.createReadStream(blob)
 
       // Extract the 'mime' property of the metadata and use it for the response header if found
       if (metadata && 'mime' in metadata && typeof metadata.mime === 'string') {
-        res.setHeader('Content-Type', metadata.mime)
+        reply.header('Content-Type', metadata.mime)
       } else {
         // Attempt to guess the MIME type based on the blob contents
         const blobSlice = await blobs.get(entry.value.blob, {
@@ -67,77 +92,39 @@ export class BlobServer {
           length: 20,
         })
 
-        const [guessedMime] = filetypemime(Array.from(blobSlice))
+        if (!blobSlice) {
+          reply.code(404)
+          throw new Error('Blob not found')
+        }
 
-        res.setHeader('Content-Type', guessedMime || 'application/octet-stream')
+        const [guessedMime] = filetypemime(blobSlice)
+
+        reply.header('Content-Type', guessedMime || 'application/octet-stream')
       }
 
-      blobsStream.pipe(res)
-    } catch (err) {
-      this.#errorHandler(res, err)
+      return reply.send(blobsStream)
     }
-  }
-
-  /**
-   *
-   * @param {import('node:http').ServerResponse} res
-   * @param {HttpError} error
-   */
-  #errorHandler(res, error) {
-    res.statusCode = error.statusCode || 500
-    res.end()
-  }
+  )
 }
 
 /**
+ * @param {string} driveId
+ * @param {import('../types').BlobType} type
+ * @param {import('../types').BlobVariant<ParsedParams['type']>} variant
+ * @param {string} name
  *
- * @param {import('node:http').IncomingMessage} req
- * @returns {BlobId}
+ * @returns {import('../types').BlobId}
  */
-function parseRequestParams(req) {
-  assert(typeof req.url === 'string', 'Invalid Request')
-  const url = new URL(req.url, `http://${req.headers.host}`)
-  const params = url.pathname.slice(1).split('/')
-  assert(params.length === 4, 'Invalid Request')
-  const [driveId, type, variant, name] = params
-  assert(type in SUPPORTED_BLOB_VARIANTS, 'Invalid Request')
-  // @ts-ignore
-  assert(SUPPORTED_BLOB_VARIANTS[type].includes(variant), 'Invalid Request')
-  return /** @type {BlobId} */ ({
+function constructBlobId(driveId, type, variant, name) {
+  assert(
+    SUPPORTED_BLOB_VARIANTS[type].includes(/** @type {*} */ (variant)),
+    `Unsupported variant "${variant}" for ${type}`
+  )
+
+  return /** @type {import('../types').BlobId} */ ({
     driveId,
     type,
     variant,
     name,
   })
-}
-
-/**
- *
- * @param {import('node:http').Server} server
- * @param {number} [port]
- * @param {string} [hostname]
- * @returns
- */
-function listenPromise(server, port, hostname) {
-  /** @type {(err: Error) => void} */
-  let errEventHandler
-  /** @type {Promise<void>} */
-  const errEvent = new Promise((resolve, reject) => {
-    errEventHandler = (err) => {
-      reject(err)
-    }
-    server.once('error', errEventHandler)
-  })
-  /** @type {Promise<void>} */
-  const listen = new Promise((resolve) => {
-    server.listen(port, hostname, () => {
-      server.removeListener('error', errEventHandler)
-      resolve()
-    })
-  })
-
-  return Promise.race([
-    errEvent, // e.g invalid port range error is always emitted before the server listening
-    listen,
-  ])
 }
